+++ WEBServer/WEBServer/Shared/Models/DeviceProbe.cs
@@ -5,7 +5,7 @@ namespace WEBServer.Shared.Models
     public class DeviceProbe
     {
         public int ID_Device { get; set; }
-        public int Delta { get; set; }
+        public short Delta { get; set; }
         public DateTime Date { get; set; }
     }
 }
